Refactor package architectures
