+++ withaeng-api/build.gradle.kts
@@ -16,7 +16,7 @@ dependencies {
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-security")
     implementation("org.springframework.boot:spring-boot-starter-validation")
-    
+
     api("org.springframework.boot:spring-boot-starter-data-jpa")
 
     // jwt
