+++ withaeng-api/src/main/kotlin/com/travel/withaeng/controller/accompany/AccompanyController.kt
@@ -1,11 +1,10 @@
-package com.travel.withaeng.accompany.controller
+package com.travel.withaeng.controller.accompany
 
 import com.travel.withaeng.common.ApiResponse
 import com.travel.withaeng.domain.accompany.*
 import jakarta.validation.Valid
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
 
 @RestController
